Moved expressions to separate folder in 'src/', added 'and' and 'or' expressions
