+++ src/calder.ts
@@ -24,9 +24,9 @@
 
 export { default as Variable } from './variable';
 export { default as SyntaxNode } from './syntaxnode';
-export { default as Expression } from './expression';
+export { default as Expression } from './expressions/expression';
 export { default as Statement } from './statement';
-export { default as Assignment } from './assignment';
+export { default as Assignment } from './expressions/assignment';
 export { default as Function } from './function';
 export { default as Reference } from './reference';
 export { default as Shader } from './shader';
