+++ src/expressions/expression.ts
@@ -1,5 +1,5 @@
-import Type from './type';
-import SyntaxNode from './syntaxnode';
+import Type from '../type';
+import SyntaxNode from '../syntaxnode';
 
 export default interface Expression extends SyntaxNode {
     returnType(): Type;
