+++ src/reference.ts
@@ -1,4 +1,4 @@
-import Expression from './expression';
+import Expression from './expressions/expression';
 import InterfaceVariable from './interface';
 import Set from './util/set';
 import Type from './type';
